feature/72 refactor: SearchResultPage search 패키지로 이동 및 글작성버튼 삭제
